+++ app/model/api/PayAndTax.scala
@@ -17,10 +17,10 @@
 package model.api
 
 import java.util.UUID
-
-import org.joda.time.LocalDate
 import play.api.libs.json.Json
 import utils.LocalDateFormat
+
+import java.time.LocalDate
 
 case class PayAndTax(payAndTaxId: UUID = UUID.randomUUID(),
                      taxablePayTotal: Option[BigDecimal],
@@ -41,4 +41,3 @@ case class PayAndTax(payAndTaxId: UUID = UUID.randomUUID(),
 object PayAndTax extends LocalDateFormat {
   implicit val formats = Json.format[PayAndTax]
 }
-
