+++ app/utils/ControllerUtils.scala
@@ -18,11 +18,8 @@ package utils
 
 import model.api.{Employment, EmploymentStatus}
 import play.api.i18n.Messages
-import uk.gov.hmrc.play.language.LanguageUtils
 
-import java.time.LocalDate
-
-object ControllerUtils extends DateHelper {
+object ControllerUtils {
 
   def getEmploymentStatus(employment: Employment)(implicit messages: Messages): String = {
     val current = Messages("lbl.employment.status.current")
@@ -33,21 +30,6 @@ object ControllerUtils extends DateHelper {
       case EmploymentStatus.Live    => current
       case EmploymentStatus.Unknown => unknown
       case _                        => ceased
-    }
-  }
-
-  private def noRecord(implicit messages: Messages): String = Messages("lbl.date.no-record")
-
-  def getStartDate(employment: Employment)(implicit messages: Messages): String = {
-    employment.startDate.fold(noRecord){date => format(date)}
-  }
-
-  def getEndDate(employment: Employment)(implicit messages: Messages): String = {
-    val ongoing = Messages("lbl.end-date.ongoing")
-    employment.employmentStatus match {
-      case EmploymentStatus.PotentiallyCeased => noRecord
-      case EmploymentStatus.Unknown           => employment.endDate.fold(noRecord){date => format(date)}
-      case _                                  => employment.endDate.fold(ongoing){date => format(date)}
     }
   }
