+++ app/views/taxhistory/employment_detail.scala.html
@@ -15,6 +15,7 @@
  *@
 
 @import model.api.{PayAndTax, CompanyBenefit, Employment, IncomeSource}
+@import utils.{Currency, ControllerUtils}
 @import uk.gov.hmrc.time.TaxYear
 @import config.AppConfig
 @import views.models.PageTitle
@@ -26,7 +27,9 @@
   employment: Employment,
   companyBenefits: List[CompanyBenefit],
   clientNameOrNino: String,
-  incomeSource: Option[IncomeSource])(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
+  incomeSource: Option[IncomeSource],
+  startDate: String,
+  endDate: String)(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
 
 @isCurrentYear = @{
     TaxYear.current.currentYear == taxYear
@@ -87,11 +90,11 @@
         }
         <div class="govuk-summary-list__row">
             <dt class="govuk-body-s govuk-!-font-weight-bold no-bottom-margin">@messages("lbl.date.start")</dt>
-            <dd class="govuk-body-s margin-inline-start">@ControllerUtils.getStartDate(employment)</dd>
+            <dd class="govuk-body-s margin-inline-start">@startDate</dd>
         </div>
         <div class="govuk-summary-list__row">
             <dt class="govuk-body-s govuk-!-font-weight-bold no-bottom-margin">@messages("lbl.date.end")</dt>
-            <dd class="govuk-body-s margin-inline-start">@ControllerUtils.getEndDate(employment)</dd>
+            <dd class="govuk-body-s margin-inline-start">@endDate</dd>
         </div>
         @if(!employment.isOccupationalPension) {
             <div class="govuk-summary-list__row">
@@ -124,7 +127,7 @@
             </table>
         }
         case None => {
-            <p id="no-payments" class="govuk-body">@messages("employmenthistory.no.pay.and.tax", employment.employerName, ControllerUtils.getStartDate(employment))</p>
+            <p id="no-payments" class="govuk-body">@messages("employmenthistory.no.pay.and.tax", employment.employerName, startDate)</p>
         }
     }
 }
